Adjusted book logic, swapped all Longs to int for mem purposes, experiments with Long were over. Added OrderService, simple order api, and a basic microservice communication and object retrieval using resttemplate to pull book from bookservice to ordder service

+++ bookservice/src/main/java/com/example/bookservice/model/Book.java
@@ -11,9 +11,7 @@ import lombok.Data;
 public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     private String title;
     private String author;
-
 }
